requirements.txt
# requirements.txt

# Core ML
transformers==4.40.1
datasets==2.19.1
evaluate==0.4.1
torch==2.2.2
accelerate==0.29.3
sentencepiece==0.2.0
rouge_score==0.1.2
py7zr==0.21.0

# Data
pandas==2.2.2
numpy==1.26.4

# Config
PyYAML==6.0.1
python-box==7.1.1
ensure==1.0.4

# HF Hub
huggingface_hub==0.22.2

# UI
streamlit==1.33.0

# Utilities
tqdm==4.66.2
python-dotenv==1.0.1